+++ packages/backend/src/server/api/endpoints/i/2fa/register-key.ts
@@ -109,13 +109,13 @@ export const meta = {
 							items: {
 								type: 'string',
 								enum: [
-									"ble",
-									"cable",
-									"hybrid",
-									"internal",
-									"nfc",
-									"smart-card",
-									"usb",
+									'ble',
+									'cable',
+									'hybrid',
+									'internal',
+									'nfc',
+									'smart-card',
+									'usb',
 								],
 							},
 						},
@@ -129,8 +129,8 @@ export const meta = {
 					authenticatorAttachment: {
 						type: 'string',
 						enum: [
-							"cross-platform",
-							"platform",
+							'cross-platform',
+							'platform',
 						],
 					},
 					requireResidentKey: {
@@ -139,9 +139,9 @@ export const meta = {
 					userVerification: {
 						type: 'string',
 						enum: [
-							"discouraged",
-							"preferred",
-							"required",
+							'discouraged',
+							'preferred',
+							'required',
 						],
 					},
 				},
@@ -150,10 +150,11 @@ export const meta = {
 				type: 'string',
 				nullable: true,
 				enum: [
-					"direct",
-					"enterprise",
-					"indirect",
-					"none",
+					'direct',
+					'enterprise',
+					'indirect',
+					'none',
+					null,
 				],
 			},
 			extensions: {