+++ packages/misskey-js/src/autogen/apiClientJSDoc.ts
@@ -1,6 +1,6 @@
 /*
  * version: 2023.12.2
- * generatedAt: 2024-01-11T14:29:04.817Z
+ * generatedAt: 2024-01-13T04:31:38.782Z
  */
 
 import type { SwitchCaseResponseType } from '../api.js';