+++ packages/misskey-js/src/autogen/models.ts
@@ -1,6 +1,6 @@
 /*
  * version: 2023.12.2
- * generatedAt: 2024-01-11T14:29:04.810Z
+ * generatedAt: 2024-01-13T04:31:38.773Z
  */
 
 import { components } from './types.js';